Add client API to get secrets
