+++ pomace/__init__.py
@@ -1,8 +1,6 @@
 from pkg_resources import DistributionNotFound, get_distribution
 
-from .api import visit
-from .config import settings
-from .models import Page, autopage
+from .api import get_secret, visit
 
 
 try:
