+++ pomace/api.py
@@ -1,7 +1,10 @@
 import atexit
 
-from . import models, utils
+from bullet import Input
+
+from . import models, shared, utils
 from .config import settings
+from .types import URL
 
 
 def visit(url: str, *, browser: str = '') -> models.Page:
@@ -14,3 +17,11 @@ def visit(url: str, *, browser: str = '') -> models.Page:
     utils.launch_browser()
 
     return models.autopage()
+
+
+def get_secret(name: str) -> str:
+    domain = URL(shared.browser.url).domain
+    cli = Input(prompt=f"{domain} {name}: ")
+    value = settings.get_secret(name) or cli.launch()
+    settings.set_secret(name, value)
+    return value
