+++ pomace/tests/test_models.py
@@ -3,12 +3,7 @@
 import pytest
 
 from ..config import settings
-from ..models import URL, Action, Locator, Page
-
-
-@pytest.fixture
-def url():
-    return URL('http://www.example.com/foo/bar')
+from ..models import Action, Locator, Page
 
 
 @pytest.fixture
@@ -24,36 +19,6 @@ def action():
 @pytest.fixture
 def page():
     return Page('example.com', '/foo/bar')
-
-
-def describe_url():
-    def describe_init():
-        def it_removes_escaped_slashes(expect, monkeypatch):
-            monkeypatch.setattr(URL, 'SLASH', '@')
-            url = URL('www.example.com', '@foo@bar')
-            expect(url.domain) == 'www.example.com'
-            expect(url.path) == '/foo/bar'
-
-    def describe_str():
-        def it_returns_the_value(expect, url):
-            expect(str(url)) == 'http://www.example.com/foo/bar'
-
-    def describe_eq():
-        def it_compares_domain_and_path(expect, url):
-            expect(url) == URL('https://www.example.com/foo/bar/')
-            expect(url) != URL('http://example.com/foo/bar')
-
-        def it_matches_patterns(expect):
-            pattern = URL('http://example.com/p/{name}')
-            expect(pattern) == URL('http://example.com/p/foobar')
-            expect(pattern) != URL('http://example.com/')
-            expect(pattern) != URL('http://example.com/p/foo/bar')
-
-    def describe_path_encoded():
-        def it_replaces_slashes_with_special_character(expect, monkeypatch, url):
-            monkeypatch.setattr(URL, 'SLASH', '@')
-            expect(url.domain) == 'www.example.com'
-            expect(url.path_encoded) == '@foo@bar'
 
 
 def describe_locator():
@@ -129,3 +94,7 @@ def describe_page():
         def it_rejects_missing_attributes(expect, page):
             with expect.raises(AttributeError):
                 getattr(page, 'foobar')
+
+    def describe_contains():
+        def it_matches_partial_html(expect, page, mockbrowser):
+            expect(page).contains("world")
