+++ pytest.ini
@@ -6,7 +6,6 @@ addopts =
 
     -r sxX
 
-    --cov=pomace
     --cov-report=html
     --cov-report=term-missing:skip-covered
     --no-cov-on-fail
