+++ scent.py
@@ -4,7 +4,6 @@ import time
 import subprocess
 
 from sniffer.api import select_runnable, file_validator, runnable
-
 try:
     from pync import Notifier
 except ImportError:
